% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metroglyphGrob.R
\name{metroglyphGrob}
\alias{metroglyphGrob}
\title{Draw a Metroglyph}
\usage{
metroglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 1,
  circle.size = 10,
  col = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  angle.start = 0,
  angle.stop = 2 * base::pi
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the distance of star gylph points from
the center.}

\item{size}{The size of rays.}

\item{circle.size}{The size of the central circle.}

\item{col}{The colour of whisker and contours.}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{angle.start}{The start angle for the glyph rays in radians. Default is zero.}

\item{angle.stop}{The stop anlge for the glyph rays in radians. Default is
\eqn{2\pi}.}
}
\value{
A \code{\link[grid]{grobTree}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a metroglyph
\insertCite{anderson_semigraphical_1957,dutoit_graphical_1986}{gglyph}.
}
\examples{
mglyph1 <- metroglyphGrob(x = 200, y = 100,
                             z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                             size = 100, circle.size = 10)

mglyph2 <- metroglyphGrob(x = 500, y = 100,
                             z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                             size = 100, circle.size = 25)

mglyph3 <- metroglyphGrob(x = 200, y = 300,
                          z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                          size = 100, circle.size = 0,
                          angle.start = base::pi, angle.stop = -base::pi)

mglyph4 <- metroglyphGrob(x = 500, y = 300,
                          z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                          size = 100, circle.size = 50,
                          angle.start = base::pi, angle.stop = -base::pi)


grid::grid.newpage()
grid::grid.draw(mglyph1)
grid::grid.draw(mglyph2)
grid::grid.draw(mglyph3)
grid::grid.draw(mglyph4)
}
\seealso{
\code{\link[gglyph]{geom_metroglyph}}
}
