% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieglyphGrob.R
\name{pieglyphGrob}
\alias{pieglyphGrob}
\title{Draw a Pie Glyph}
\usage{
pieglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 1,
  edges = 200,
  col = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  angle.start = 0,
  angle.stop = 2 * base::pi,
  linejoin = c("mitre", "round", "bevel"),
  scale.segment = FALSE,
  scale.radius = TRUE
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the values to be plotted as dimensions
of the pie glyph according to the arguments \code{scale.segment} or
\code{scale.radius}.}

\item{size}{The size of glyphs.}

\item{edges}{The number of edges of the polygon to depict the circular glyph
outline.}

\item{col}{The line colour.}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{angle.start}{The start angle for the glyph in radians. Default is zero.}

\item{angle.stop}{The stop anlge for the glyph in radians. Default is
\eqn{2\pi}.}

\item{linejoin}{The line join style for the pie segment polygons. Either
\code{"mitre"}, \code{"round"} or \code{"bevel"}.}

\item{scale.segment}{logical. If \code{TRUE}, the segments (pie slices) are
scaled according to value of \code{z}.}

\item{scale.radius}{logical. If \code{TRUE}, the radius of segments (pie
slices) are scaled according to value of \code{z}.}
}
\value{
A \code{\link[grid]{grobTree}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a circular pie or clock glyph
\insertCite{ward_visualization_2000,fuchs_evaluation_2013}{gglyph}.
}
\examples{
p1 <- pieglyphGrob(x = 150, y = 150,
             z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
             size = 50)

p2 <- pieglyphGrob(x = 300, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.radius = FALSE)

p3 <- pieglyphGrob(x = 450, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.segment = TRUE, scale.radius = FALSE)

p4 <- pieglyphGrob(x = 150, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, angle.start = 0, angle.stop = -base::pi)

p5 <- pieglyphGrob(x = 300, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.radius = FALSE,
                   angle.start = 0, angle.stop = -base::pi)

p6 <- pieglyphGrob(x = 450, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.segment = TRUE, scale.radius = FALSE,
                   angle.start = 0, angle.stop = -base::pi)

grid::grid.newpage()
grid::grid.draw(p1)
grid::grid.draw(p2)
grid::grid.draw(p3)
grid::grid.draw(p4)
grid::grid.draw(p5)
grid::grid.draw(p6)

p1 <- pieglyphGrob(x = 150, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, fill = RColorBrewer::brewer.pal(6, "Dark2"))

p2 <- pieglyphGrob(x = 300, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.radius = FALSE,
                   fill = RColorBrewer::brewer.pal(6, "Dark2"))

p3 <- pieglyphGrob(x = 450, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.segment = TRUE, scale.radius = FALSE,
                   fill = RColorBrewer::brewer.pal(6, "Dark2"))

p4 <- pieglyphGrob(x = 150, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, angle.start = 0, angle.stop = -base::pi,
                   fill = RColorBrewer::brewer.pal(6, "Dark2"))

p5 <- pieglyphGrob(x = 300, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.radius = FALSE,
                   angle.start = 0, angle.stop = -base::pi,
                   fill = RColorBrewer::brewer.pal(6, "Dark2"))

p6 <- pieglyphGrob(x = 450, y = 350,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 50, scale.segment = TRUE, scale.radius = FALSE,
                   angle.start = 0, angle.stop = -base::pi,
                   fill = RColorBrewer::brewer.pal(6, "Dark2"))

grid::grid.newpage()
grid::grid.draw(p1)
grid::grid.draw(p2)
grid::grid.draw(p3)
grid::grid.draw(p4)
grid::grid.draw(p5)
grid::grid.draw(p6)

p1 <- pieglyphGrob(x = 200, y = 150,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 100, lwd = 5)

p2 <- pieglyphGrob(x = 400, y = 300,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 100, lwd = 5, linejoin = "round")

p3 <- pieglyphGrob(x = 600, y = 450,
                   z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33),
                   size = 100, lwd = 5, linejoin = "bevel")

grid::grid.newpage()
grid::grid.draw(p1)
grid::grid.draw(p2)
grid::grid.draw(p3)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{geom_pieglyph}}
}
