% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa2022_subset}
\alias{pisa2022_subset}
\title{PISA 2022 data}
\format{
A data frame with 189,966 rows and 3 variables:
\describe{
\item{\code{CNT}}{Country of the respondent (character).}
\item{\code{HISCED}}{Highest educational level (ISCED) attained by parents of the respondent (integer).}
\item{\code{MATH}}{Math score of the respondent (numeric).}
}
}
\source{
Data obtained from:
\itemize{
\item OECD (2023). \emph{PISA 2022 Database} [Data Set]. Zenodo. \url{https://doi.org/10.5281/zenodo.13382904}
}
}
\usage{
pisa2022_subset
}
\description{
A subset from the PISA 2022 data.
}
\examples{
data(pisa2022_subset)
head(pisa2022_subset)
}
\references{
Additional references for further reading:
\itemize{
\item OECD (2024). \emph{PISA 2022 Technical Report}. OECD Publishing, Paris. \url{https://doi.org/10.1787/01820d6d-en}
\item OECD (2024). \emph{Programme for International Student Assessment (PISA)}. \url{https://www.oecd.org/en/about/programmes/pisa.html}
}
}
\keyword{datasets}
