% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pisa_2022}
\alias{pisa_2022}
\title{PISA 2022 data}
\format{
A data frame with 492 rows and 3 variables:
\describe{
\item{\code{from}}{Category A of educational level (ISCED) attained by the parents of the respondent (character).}
\item{\code{to}}{Category B of educational level (ISCED) attained by the parents of the respondent (character).}
\item{\code{group}}{Country of the respondent (character).}
\item{\code{sig}}{p-value of the pairwise t-test (numeric).}
}
}
\source{
Data obtained from:
\itemize{
\item OECD (2023). \emph{PISA 2022 Database} [Data Set]. Zenodo. \doi{10.5281/zenodo.13382904}
}
Licensed under CC BY 4.0: \url{https://creativecommons.org/licenses/by/4.0/}
}
\usage{
data(pisa_2022)
}
\description{
Results of pairwise t-tests (with Bonferroni correction) performed on a subset from the PISA 2022 data.
}
\examples{
data(pisa_2022)
head(pisa_2022)
}
\references{
Additional reference(s) for further reading:
\itemize{
\item OECD (2024). \emph{PISA 2022 Technical Report}. OECD Publishing, Paris. \doi{10.1787/01820d6d-en}
}
}
\keyword{dataset}
