% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_z_continuous}
\alias{scale_z_continuous}
\title{Alter scales for continuous data mapped to glyphs}
\usage{
scale_z_continuous(..., range = c(0.1, 1), z)
}
\arguments{
\item{...}{Additional arguments to be passed on to \code{\link[ggplot2]{continuous_scale}}.}

\item{range}{The range to which the variable(s) specified in argument \code{z} are to be scaled.}

\item{z}{The variable(s) mapped to the glyph as a character vector.}
}
\description{
Scale variable(s) mapped to the glyphs to the given
range using \code{\link[scales]{rescale_pal}}.
}
