% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipri_milex1949_2023_subset}
\alias{sipri_milex1949_2023_subset}
\title{SIPRI Military Expenditure 1949-2023 data}
\format{
A data frame with 48 rows and 3 variables:
\describe{
\item{\code{Country}}{Country name (character).}
\item{\code{Year}}{Year (integer).}
\item{\code{Spending}}{Military spending in current US dollars (numeric).}
}
}
\source{
Data obtained from:
\itemize{
\item SIPRI (2025). \emph{SIPRI Military Expenditure Database} [Data Set]. \url{https://doi.org/10.55163/CQGC9685}
}
}
\usage{
sipri_milex1949_2023_subset
}
\description{
A subset of the SIPRI Military Expenditure 1949-2023 data.
}
\examples{
data(sipri_milex1949_2023_subset)
head(sipri_milex1949_2023_subset)
}
\keyword{datasets}
