% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipri_milex_1995_2023}
\alias{sipri_milex_1995_2023}
\title{SIPRI Military Expenditure 1995-2023}
\format{
A data frame with 77 rows and 3 variables:
\describe{
\item{\code{from}}{Name of country A (character).}
\item{\code{to}}{Name of country B (character).}
\item{\code{group}}{Year (numeric).}
}
}
\source{
Data obtained from:
\itemize{
\item SIPRI (2025). \emph{SIPRI Military Expenditure Database} [Data Set]. \doi{10.55163/CQGC9685}
}
Licensed under CC BY 4.0: \url{https://creativecommons.org/licenses/by/4.0/}
}
\usage{
data(sipri_milex_1995_2023)
}
\description{
A subset of the SIPRI Military Expenditure 1949-2023 data.
}
\examples{
data(sipri_milex_1995_2023)
head(sipri_milex_1995_2023)
}
\keyword{dataset}
