% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tileglyphGrob.R
\name{tileglyphGrob}
\alias{tileglyphGrob}
\title{Draw a Tile Glyph}
\usage{
tileglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 10,
  ratio = 1,
  nrow = 1,
  col = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  linejoin = c("mitre", "round", "bevel")
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the values to be plotted as dimensions
in the tileglyph.}

\item{size}{The size of glyphs.}

\item{ratio}{The aspect ratio (height / width).}

\item{nrow}{The number of rows.}

\item{col}{The line colour.}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{linejoin}{The line join style for the tile polygon. Either
\code{"mitre"}, \code{"round"} or \code{"bevel"}.}
}
\value{
A \code{\link[grid]{grob}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a tile glyph similar to
'autoglyph' \insertCite{beddow_shape_1990}{gglyph} or 'stripe glyph'
\insertCite{fuchs_evaluation_2013}{gglyph}.
}
\examples{
tg1 <- tileglyphGrob(x = 150, y = 150,
                   z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                   size = 30)

tg2 <- tileglyphGrob(x = 450, y = 150,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 30)

tg3 <- tileglyphGrob(x = 150, y = 250,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30, nrow = 2)

tg4 <- tileglyphGrob(x = 450, y = 250,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 30, nrow = 2)

tg5 <- tileglyphGrob(x = 150, y = 350,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg6 <- tileglyphGrob(x = 450, y = 350,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 30,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg7 <- tileglyphGrob(x = 150, y = 450,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30, nrow = 2,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg8 <- tileglyphGrob(x = 450, y = 450,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 30, nrow = 2,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

grid::grid.newpage()
grid::grid.draw(tg1)
grid::grid.draw(tg2)
grid::grid.draw(tg3)
grid::grid.draw(tg4)
grid::grid.draw(tg5)
grid::grid.draw(tg6)
grid::grid.draw(tg7)
grid::grid.draw(tg8)

tg1 <- tileglyphGrob(x = 150, y = 150,
                   z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                   size = 5, ratio = 10)

tg2 <- tileglyphGrob(x = 450, y = 150,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 5, ratio = 10)

tg3 <- tileglyphGrob(x = 150, y = 250,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 5, nrow = 2, ratio = 10)

tg4 <- tileglyphGrob(x = 450, y = 250,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 5, nrow = 2, ratio = 10)

tg5 <- tileglyphGrob(x = 150, y = 350,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 5, ratio = 10,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg6 <- tileglyphGrob(x = 450, y = 350,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 5, ratio = 10,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg7 <- tileglyphGrob(x = 150, y = 450,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 5, nrow = 2, ratio = 10,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

tg8 <- tileglyphGrob(x = 450, y = 450,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7),
                     size = 5, nrow = 2, ratio = 10,
                     fill = RColorBrewer::brewer.pal(7, "Dark2"))

grid::grid.newpage()
grid::grid.draw(tg1)
grid::grid.draw(tg2)
grid::grid.draw(tg3)
grid::grid.draw(tg4)
grid::grid.draw(tg5)
grid::grid.draw(tg6)
grid::grid.draw(tg7)
grid::grid.draw(tg8)

tg1 <- tileglyphGrob(x = 150, y = 150,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30, nrow = 2, lwd = 5)

tg2 <- tileglyphGrob(x = 300, y = 300,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30, nrow = 2, lwd = 5,
                     linejoin = "round")

tg3 <- tileglyphGrob(x = 450, y = 450,
                     z = c(4, 3.5, 2.7, 6.8, 3.4, 5.7, 4.3),
                     size = 30, nrow = 2, lwd = 5,
                     linejoin = "bevel")


grid::grid.newpage()
grid::grid.draw(tg1)
grid::grid.draw(tg2)
grid::grid.draw(tg3)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{geom_tileglyph}}
}
