% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_starglyph.R
\encoding{UTF-8}
\name{geom_starglyph}
\alias{geom_starglyph}
\title{Add Star Glyphs as a Scatterplot}
\usage{
geom_starglyph(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  cols = character(0L),
  whisker = TRUE,
  contour = TRUE,
  colour.whisker = NULL,
  colour.contour = NULL,
  colour.points = NULL,
  linewidth.whisker = 1,
  linewidth.contour = 1,
  full = TRUE,
  draw.grid = FALSE,
  grid.point.size = 1,
  show.legend = NA,
  repel = FALSE,
  repel.control = gglyph.repel.control(),
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{layer}()}. These
are often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "green"} or \code{size = 3}. They may also be parameters to
the paired geom/stat.}

\item{cols}{Name of columns specifying the variables to be plotted in the
glyphs as a character vector.}

\item{whisker}{logical. If \code{TRUE}, plots the star glyph whiskers.}

\item{contour}{logical. If \code{TRUE}, plots the star glyph contours.}

\item{colour.whisker}{The colour of whiskers.}

\item{colour.contour}{The colour of contours.}

\item{colour.points}{The colour of grid points.}

\item{linewidth.whisker}{The whisker line width.}

\item{linewidth.contour}{The contour line width.}

\item{full}{logical. If \code{TRUE}, full star glyphs (360°) are plotted,
otherwise half star glyphs (180°) are plotted.}

\item{draw.grid}{logical. If \code{TRUE}, grid points are plotted along the
whiskers if all the variables specified in \code{cols} are of type
\link[base]{factor}. Default is \code{FALSE}.}

\item{grid.point.size}{The size of the grid points in native units.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{repel}{logical. If \code{TRUE}, the glyphs are repel away from each
other to avoid overlaps. Default is \code{FALSE}.}

\item{repel.control}{A list of control settings for the repel algorithm.
Ignored if \code{repel = FALSE}. See
\code{\link[gglyph]{gglyph.repel.control}} for details  on the various
control parameters.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \code{geom} layer.
}
\description{
The starglyph geom is used to plot multivariate data as star glyphs
\insertCite{siegel_surgical_1972,chambers_graphical_1983,dutoit_graphical_1986}{gglyph}
in a scatterplot.
}
\section{Aesthetics}{
 \code{geom_starglyph()} understands the following
  aesthetics (required aesthetics are in bold): \itemize{ \item{\strong{x}}
  \item{\strong{y}} \item{alpha} \item{colour} \item{fill} \item{group}
  \item{shape} \item{size} \item{stroke} \item{linetype} }

  See \code{vignette("ggplot2-specs", package = "ggplot2")} for further
  details on setting these aesthetics.

  The following additional aesthetics are considered if \code{repel = TRUE}:
  \itemize{ \item{point.size} \item{segment.linetype} \item{segment.colour}
  \item{segment.size} \item{segment.alpha} \item{segment.curvature}
  \item{segment.angle} \item{segment.ncp} \item{segment.shape}
  \item{segment.square} \item{segment.squareShape} \item{segment.inflect}
  \item{segment.debug} }

  See \code{ggrepel}
  \href{https://ggrepel.slowkow.com/articles/examples.html}{examples} page
  for further details on setting these aesthetics.
}

\examples{
# Scale the data
zs <- c("hp", "drat", "wt", "qsec", "vs", "am", "gear", "carb")
mtcars[ , zs] <- lapply(mtcars[ , zs], scales::rescale)

mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$lab <- row.names(mtcars)

library(ggplot2)
theme_set(theme_bw())
options(ggplot2.discrete.colour = RColorBrewer::brewer.pal(8, "Dark2"))
options(ggplot2.discrete.fill = RColorBrewer::brewer.pal(8, "Dark2"))

# Both whiskers and contour
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha =  0.5) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha =  0.5, full = FALSE) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha =  0.5,
                 linewidth.whisker = 3, linewidth.contour = 0.1) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha =  0.5,
                 linewidth.whisker = 1, linewidth.contour = 3) +
  ylim(c(-0, 550))


# Only contours (polygon)
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = FALSE, contour = TRUE,
                 size = 10, alpha =  0.5) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = FALSE, contour = TRUE,
                 size = 10, alpha =  0.5, linewidth.contour = 3) +
  ylim(c(-0, 550))

# Only whiskers
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, colour = cyl),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 10) +
  geom_point(data = mtcars, aes(x = mpg, y = disp, colour = cyl)) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, colour = cyl),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 10, full = FALSE) +
  geom_point(data = mtcars, aes(x = mpg, y = disp, colour = cyl)) +
  ylim(c(-0, 550))

# Whiskers with colours
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 10,
                 colour.whisker = RColorBrewer::brewer.pal(8, "Dark2")) +
  geom_point(data = mtcars, aes(x = mpg, y = disp)) +
  ylim(c(-0, 550))

# With text annotations
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, colour = cyl),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 10) +
  geom_point(data = mtcars, aes(x = mpg, y = disp, colour = cyl)) +
  geom_text(data = mtcars, aes(x = mpg, y = disp, label = lab), cex = 2) +
  ylim(c(-0, 550))

# Faceted
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha =  0.5) +
  ylim(c(-0, 550)) +
  facet_grid(. ~ cyl)

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, colour = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10) +
  ylim(c(-0, 550)) +
  facet_grid(. ~ cyl)

# Repel glyphs
ggplot(data = mtcars) +
  geom_point(aes(x = mpg, y = disp, colour = cyl)) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 10, alpha = 1, repel = TRUE) +
  ylim(c(-0, 550)) +
  xlim(c(8, 35))

rm(mtcars)
mtcars[ , zs] <- lapply(mtcars[ , zs], scales::rescale)

mtcars[ , zs] <- lapply(mtcars[, zs],
                        function(x) cut(x, breaks = 3,
                                        labels = c(1, 2, 3)))
mtcars[ , zs] <- lapply(mtcars[ , zs], as.factor)

mtcars$cyl <- as.factor(mtcars$cyl)
mtcars$lab <- row.names(mtcars)

# Grid points
ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, fill = cyl),
                 cols = zs, whisker = TRUE, contour = TRUE,
                 size = 3, alpha =  0.5, draw.grid = TRUE,
                 grid.point.size = 5) +
  ylim(c(-0, 550))


ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp, colour = cyl),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 3, draw.grid = TRUE, grid.point.size = 7,
                 linewidth.whisker = 2, alpha = 0.7) +
  ylim(c(-0, 550))

ggplot(data = mtcars) +
  geom_starglyph(aes(x = mpg, y = disp),
                 cols = zs, whisker = TRUE, contour = FALSE,
                 size = 3, draw.grid = TRUE,
                 grid.point.size = 5, alpha =  0.8,
                 colour.whisker = RColorBrewer::brewer.pal(8, "Dark2")) +
  geom_point(data = mtcars, aes(x = mpg, y = disp)) +
  ylim(c(-0, 550))

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{starglyphGrob}}

Other geoms: 
\code{\link{geom_dotglyph}()},
\code{\link{geom_metroglyph}()},
\code{\link{geom_pieglyph}()},
\code{\link{geom_profileglyph}()},
\code{\link{geom_tileglyph}()}
}
\concept{geoms}
