% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.repel.control.R
\encoding{UTF-8}
\name{gglyph.repel.control}
\alias{gglyph.repel.control}
\title{Control parameters for the repel algorithm}
\usage{
gglyph.repel.control(
  box.padding = 0.25,
  point.padding = 1e-06,
  min.segment.length = 0.5,
  arrow = NULL,
  force = 1,
  force_pull = 1,
  max.time = 0.5,
  max.iter = 10000,
  max.overlaps = getOption("ggrepel.max.overlaps", default = 10),
  nudge_x = 0,
  nudge_y = 0,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  na.rm = FALSE,
  direction = c("both", "y", "x"),
  seed = NA,
  verbose = FALSE
)
}
\arguments{
\item{box.padding}{Amount of padding around bounding box, as unit or number.
Defaults to 0.25. (Default unit is lines, but other units can be specified
by passing \code{unit(x, "units")}).}

\item{point.padding}{Amount of padding around labeled point, as unit or
number. Defaults to 0. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{min.segment.length}{Skip drawing segments shorter than this, as unit or
number. Defaults to 0.5. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{arrow}{specification for arrow heads, as created by
\code{\link[grid]{arrow}}}

\item{force}{Force of repulsion between overlapping text labels. Defaults to
1.}

\item{force_pull}{Force of attraction between a text label and its
corresponding data point. Defaults to 1.}

\item{max.time}{Maximum number of seconds to try to resolve overlaps.
Defaults to 0.5.}

\item{max.iter}{Maximum number of iterations to try to resolve overlaps.
Defaults to 10000.}

\item{max.overlaps}{Exclude text labels that overlap too many things.
Defaults to 10.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustments to nudge the
starting position of each text label. The units for \code{nudge_x} and
\code{nudge_y} are the same as for the data units on the x-axis and y-axis.}

\item{xlim, ylim}{Limits for the x and y axes. Text labels will be constrained
to these limits. By default, text labels are constrained to the entire plot
area.}

\item{direction}{"both", "x", or "y" -- direction in which to adjust position
of labels}

\item{seed}{Random seed passed to \code{\link[base]{set.seed}}. Defaults to
\code{NA}, which means that \code{set.seed} will not be called.}

\item{verbose}{If \code{TRUE}, some diagnostics of the repel algorithm are
printed}

\item{repel}{logical. If \code{TRUE}, the glyphs are repel away from each
other to avoid overlaps. Default is \code{FALSE}.}

\item{repel.control}{A list of control settings for the repel algorithm.
Ignored if \code{repel = FALSE}. See
\code{\link[gglyph]{gglyph.repel.control}} for details  on the various
control parameters.}
}
\value{
A list with the following components to control the repel algorithm
  corresponding to the same in \strong{Arguments}. \item{box.padding}{}
  \item{point.padding}{} \item{min.segment.length}{} \item{arrow}{}
  \item{force}{} \item{force_pull}{} \item{max.time}{} \item{max.iter}{}
  \item{max.overlaps}{} \item{nudge_x, nudge_y}{} \item{xlim, ylim}{}
  \item{direction}{} \item{seed}{} \item{verbose}{} \item{repel}{}
  \item{repel.control}{}
}
\description{
Set the contol parameters for the repel algorithm
\insertCite{slowikowski_ggrepel_2021}{gglyph} used in various geoms
(\code{geom_*()}) implemented in \code{gglyph} to repel the glyphs from each
other.
}
\examples{

# Adjust force
gglyph.repel.control(force = 0.5)

# Adjust max.iter
gglyph.repel.control(max.iter = 5000)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[ggrepel]{ggrepel}}
}
