% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileglyphGrob.R
\name{profileglyphGrob}
\alias{profileglyphGrob}
\title{Draw a Profile Glyph}
\usage{
profileglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 1,
  col.bar = "black",
  col.line = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  width = 10,
  flip.axes = FALSE,
  bar = TRUE,
  line = TRUE,
  mirror = TRUE,
  linejoin = c("mitre", "round", "bevel")
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the values to be plotted as dimensions
of the profile (length of the bars).}

\item{size}{The size of glyphs.}

\item{col.bar}{The colour of bars.}

\item{col.line}{The colour of line(s).}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{width}{The width of the bars.}

\item{flip.axes}{logical. If \code{TRUE}, axes are flipped.}

\item{bar}{logical. If \code{TRUE}, profile bars are plotted.}

\item{line}{logical. If \code{TRUE}, profile line is plotted.}

\item{mirror}{logical. If \code{TRUE}, mirror profile is plotted.}

\item{linejoin}{The line join style for the profile line(s) and bars. Either
\code{"mitre"}, \code{"round"} or \code{"bevel"}.}
}
\value{
A \code{\link[grid]{grobTree}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a profile glyph
\insertCite{chambers_graphical_1983,dutoit_graphical_1986}{gglyph}.
}
\examples{
# mirror = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100,
                             col.bar = "salmon", col.line = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE,
                                    col.bar = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 col.line = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             fill = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)


# mirror = FALSE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 200, z = dims,
                             size = 100,
                             mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 200, z = dims,
                                    size = 100, line = FALSE,
                                    mirror = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 200, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 mirror = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 350, z = dims,
                             size = 100, mirror = FALSE,
                             col.bar = "salmon", col.line = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 350, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    col.bar = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 350, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 mirror = FALSE, col.line = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)#'

barglyph <- profileglyphGrob(x = 100, y = 500, z = dims, size = 100,
                             fill = "salmon", mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 500, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 500, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 mirror = FALSE, fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

# mirror = TRUE, flip.axes = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100, flip.axes = TRUE)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100, flip.axes = TRUE,
                             col.bar = "salmon", col.line = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    col.bar = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 col.line = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             flip.axes = TRUE,
                             fill = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)


# mirror = FALSE, flip.axes = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100, flip.axes = TRUE,
                             mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    mirror = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100, mirror = FALSE,
                             flip.axes = TRUE,
                             col.bar = "salmon", col.line = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    flip.axes = TRUE,
                                    col.bar = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE, col.line = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             flip.axes = TRUE,
                             fill = "salmon", mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    flip.axes = TRUE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE, fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

# linejoin variants
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
pg1 <- profileglyphGrob(x = 100, y = 150, z = dims,
                        size = 150, lwd = 5, width = 25)

pg2 <- profileglyphGrob(x = 250, y = 250, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "round")

pg3 <- profileglyphGrob(x = 400, y = 350, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "bevel")

grid::grid.newpage()
grid::grid.draw(pg1)
grid::grid.draw(pg2)
grid::grid.draw(pg3)

dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
pg1 <- profileglyphGrob(x = 100, y = 150, z = dims,
                        size = 150, lwd = 5, width = 25,
                        bar = FALSE)

pg2 <- profileglyphGrob(x = 250, y = 250, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "round", bar = FALSE)

pg3 <- profileglyphGrob(x = 400, y = 350, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "bevel", bar = FALSE)

grid::grid.newpage()
grid::grid.draw(pg1)
grid::grid.draw(pg2)
grid::grid.draw(pg3)

dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
pg1 <- profileglyphGrob(x = 100, y = 150, z = dims,
                        size = 150, lwd = 5, width = 25,
                        line = FALSE)

pg2 <- profileglyphGrob(x = 250, y = 250, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "round", line = FALSE)

pg3 <- profileglyphGrob(x = 400, y = 350, z = dims,
                        size = 150, lwd = 5, width = 25,
                        linejoin = "bevel", line = FALSE)

grid::grid.newpage()
grid::grid.draw(pg1)
grid::grid.draw(pg2)
grid::grid.draw(pg3)

bg1 <- profileglyphGrob(x = 100, y = 100, z = dims,
                        size = 100,
                        fill = RColorBrewer::brewer.pal(6, "Dark2"))

bpg1 <- profileglyphGrob(x = 300, y = 100, z = dims,
                         size = 100, line = FALSE,
                         fill = RColorBrewer::brewer.pal(6, "Dark2"))

bg2 <- profileglyphGrob(x = 150, y = 250, z = dims,
                        size = 100, mirror = FALSE,
                        fill = RColorBrewer::brewer.pal(6, "Dark2"))

bpg2 <- profileglyphGrob(x = 350, y = 250, z = dims,
                         size = 100, line = FALSE, mirror = FALSE,
                         fill = RColorBrewer::brewer.pal(6, "Dark2"))

bg3 <- profileglyphGrob(x = 100, y = 300, z = dims,
                        size = 100, flip.axes = TRUE,
                        fill = RColorBrewer::brewer.pal(6, "Dark2"))

bpg3 <- profileglyphGrob(x = 300, y = 300, z = dims,
                         size = 100, line = FALSE, flip.axes = TRUE,
                         fill = RColorBrewer::brewer.pal(6, "Dark2"))

bg4 <- profileglyphGrob(x = 150, y = 400, z = dims,
                        size = 100, mirror = FALSE, flip.axes = TRUE,
                        fill = RColorBrewer::brewer.pal(6, "Dark2"))

bpg4 <- profileglyphGrob(x = 350, y = 400, z = dims,
                         size = 100, line = FALSE, mirror = FALSE,
                         flip.axes = TRUE,
                         fill = RColorBrewer::brewer.pal(6, "Dark2"))


grid::grid.newpage()
grid::grid.draw(bg1)
grid::grid.draw(bpg1)
grid::grid.draw(bg2)
grid::grid.draw(bpg2)
grid::grid.draw(bg3)
grid::grid.draw(bpg3)
grid::grid.draw(bg4)
grid::grid.draw(bpg4)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{geom_profileglyph}}
}
