% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileglyphGrob.R
\name{profileglyphGrob}
\alias{profileglyphGrob}
\title{Draw a Profile Glyph}
\usage{
profileglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 1,
  col = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  width = 10,
  flip.axes = FALSE,
  bar = TRUE,
  line = TRUE,
  mirror = TRUE
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the distance of star gylph points from
the center.}

\item{size}{The size of glyphs.}

\item{col}{The colour of whisker and contours.}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{width}{The width of the bars.}

\item{flip.axes}{logical. If \code{TRUE}, axes are flipped.}

\item{bar}{logical. If \code{TRUE}, profile bars are plotted.}

\item{line}{logical. If \code{TRUE}, profile line is plotted.}

\item{mirror}{logical. If \code{TRUE}, mirror profile is plotted.}
}
\value{
A \code{\link[grid]{grobTree}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a profile glyph
\insertCite{chambers_graphical_1983,dutoit_graphical_1986}{gglyph}.
}
\examples{
# mirror = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100,
                             col = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE,
                                    col = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 col = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             fill = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)


# mirror = FALSE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 200, z = dims,
                             size = 100,
                             mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 200, z = dims,
                                    size = 100, line = FALSE,
                                    mirror = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 200, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 mirror = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 350, z = dims,
                             size = 100, mirror = FALSE,
                             col = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 350, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    col = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 350, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 mirror = FALSE, col = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 500, z = dims, size = 100,
                             fill = "salmon", mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 500, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 500, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 mirror = FALSE, fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)






# mirror = TRUE, flip.axes = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100, flip.axes = TRUE)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100, flip.axes = TRUE,
                             col = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    col = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 col = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             flip.axes = TRUE,
                             fill = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)


# mirror = FALSE, flip.axes = TRUE
dims = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33)
barglyph <- profileglyphGrob(x = 100, y = 100, z = dims,
                             size = 100, flip.axes = TRUE,
                             mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 100, z = dims,
                                    size = 100, line = FALSE,
                                    flip.axes = TRUE,
                                    mirror = FALSE)

profileglyph <- profileglyphGrob(x = 500, y = 100, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE)
grid::grid.newpage()
grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 250, z = dims,
                             size = 100, mirror = FALSE,
                             flip.axes = TRUE,
                             col = "salmon")

barprofileglyph <- profileglyphGrob(x = 300, y = 250, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    flip.axes = TRUE,
                                    col = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 250, z = dims,
                                 size = 100, line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE, col = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)

barglyph <- profileglyphGrob(x = 100, y = 400, z = dims, size = 100,
                             flip.axes = TRUE,
                             fill = "salmon", mirror = FALSE)

barprofileglyph <- profileglyphGrob(x = 300, y = 400, z = dims,
                                    size = 100, line = FALSE, mirror = FALSE,
                                    flip.axes = TRUE,
                                    fill = "cyan")

profileglyph <- profileglyphGrob(x = 500, y = 400, z = dims, size = 100,
                                 line = TRUE, bar = FALSE,
                                 flip.axes = TRUE,
                                 mirror = FALSE, fill = "green")

grid::grid.draw(barglyph)
grid::grid.draw(barprofileglyph)
grid::grid.draw(profileglyph)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{geom_profileglyph}}
}
