% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starglyphGrob.R
\name{starglyphGrob}
\alias{starglyphGrob}
\title{Draw a Star Glyph}
\usage{
starglyphGrob(
  x = 0.5,
  y = 0.5,
  z,
  size = 1,
  col = "black",
  fill = NA,
  lwd = 1,
  alpha = 1,
  angle.start = 0,
  angle.stop = 2 * base::pi,
  whisker = TRUE,
  contour = TRUE
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-locations.}

\item{y}{A numeric vector or unit object specifying y-locations.}

\item{z}{A numeric vector specifying the distance of star gylph points from
the center.}

\item{size}{The size of glyphs.}

\item{col}{The colour of whisker and contours.}

\item{fill}{The fill colour.}

\item{lwd}{The line width.}

\item{alpha}{The alpha transparency value.}

\item{angle.start}{The start angle for the glyph in radians. Default is zero.}

\item{angle.stop}{The stop anlge for the glyph in radians. Default is
\eqn{2\pi}.}

\item{whisker}{logical. If \code{TRUE}, plots the star glyph whiskers.}

\item{contour}{logical. If \code{TRUE}, plots the star glyph contours. glyph.}
}
\value{
A \code{\link[grid]{grobTree}} object.
}
\description{
Uses \code{\link[grid]{Grid}} graphics to draw a star glyph
\insertCite{siegel_surgical_1972,chambers_graphical_1983,dutoit_graphical_1986}{gglyph}.
}
\examples{
stargylph <- starglyphGrob(x = 250, y = 250,
                           z = c(0.24, 0.3, 0.8, 1.4, 0.6, 0.33), size = 100)
grid::grid.newpage()
grid::grid.draw(stargylph)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[gglyph]{geom_starglyph}}
}
